Fixed the type conversion with an explicit typecast

+++ num_types/src/main.rs
@@ -1,5 +1,5 @@
 fn main() {
     let mut a = 5u32;
-    a = 4i32;
+    a = 4i32 as u32;
     println!("a: {}", a);
 }
